App_Deployment/requirements.txt
gunicorn
dash
flask_login
dash_bootstrap_components
sqlalchemy
flask_sqlalchemy
python-dotenv
numpy
mysqlclient
mysql-connector-python
joblib
scikit-learn
pandas

